+++ lib/src/runner/steps/exit_app_step.dart
@@ -1,0 +1,49 @@
+import 'dart:io';
+
+import 'package:automated_testing_framework/automated_testing_framework.dart';
+import 'package:flutter/material.dart';
+import 'package:json_class/json_class.dart';
+import 'package:meta/meta.dart';
+
+/// Test step that exits the application.
+class ExitAppStep extends TestRunnerStep {
+  /// Creates an instance from a JSON-like map structure.  This expects the
+  /// following format:
+  ///
+  /// ```json
+  /// {
+  /// }
+  /// ```
+  ///
+  /// See also:
+  /// * [JsonClass.parseBool]
+  /// * [JsonClass.parseDurationFromSeconds]
+  static ExitAppStep fromDynamic(dynamic map) {
+    ExitAppStep result;
+
+    if (map != null) {
+      result = ExitAppStep();
+    }
+
+    return result;
+  }
+
+  /// Executes the step and exits the application.
+  @override
+  Future<void> execute({
+    @required TestReport report,
+    @required TestController tester,
+  }) async {
+    var name = 'exitApp()';
+    log(
+      name,
+      tester: tester,
+    );
+    exit(1);
+  }
+
+  /// Converts this to a JSON compatible map.  For a description of the format,
+  /// see [fromDynamic].
+  @override
+  Map<String, dynamic> toJson() => {};
+}